backend/carbon-intelligence/scrapers/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
yfinance==0.2.32
feedparser==6.0.10
beautifulsoup4==4.12.2
